fix(git): initialize gitcmd module before crash recovery uses git commands
The recoverIncompletePushes() function uses gitcmd.NewCommand() during
GitRepositoryService.Initialize(), but the gitea gitcmd module requires
setting.Git.HomePath to be set first. Without this, HomeDir() fails with:
'Unable to init Git's HomeDir, incorrect initialization of the setting
and git modules'

Fix: Move initGitCmd() to git_repository_service.go and call it early
in Initialize() before any git operations are attempted.
